TUI-9 major refactor, logger added, corrected logic of initial price assignment
