+++ src/main/resources/application-local.properties
@@ -6,3 +6,5 @@ spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
+
+logging.level.root=INFO
