+++ src/main/resources/application.properties
@@ -11,10 +11,10 @@ jasypt.encryptor.password=adrian
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://ec2-174-129-33-27.compute-1.amazonaws.com:5432/d7cj0f90smp2pe?sslmode=require&user=fopxwnotszeqdx&password=a3d6ba756f41f91988012e2592531e4bea206d13c3722e9786cb04b8d80c17b8
-spring.datasource.username=postgres
-spring.datasource.password=root
+spring.datasource.url=jdbc:postgresql://ec2-174-129-33-14.compute-1.amazonaws.com:5432/dceba6ffp8cnja?sslmode=require&user=fdkozowgijvzgl&password=2078aa2fd09a0b7c72368e07467f284cfa7fa989234c881a5109437ea8f5f6ff
 spring.datasource.initialization-mode=always
 spring.datasource.initialize=true
 #spring.datasource.schema=classpath:/PRICE_CHECKER.sql //for initial data insert --> can insert historic data.
 spring.datasource.continue-on-error=true
+
+logging.level.root=INFO
