+++ src/main/resources/banner.txt
@@ -1,0 +1,7 @@
+ $$$$$$$$$$$$$     $$$     $$$     $$$$               $$$$$     $$$     $$$     $$$$$$$$$$          $$$$$       $$$$    $$$$
+     $$$$$         $$$     $$$     $$$$             $$$         $$$     $$$     $$$               $$$           $$$$  $$$$
+     $$$$$         $$$     $$$     $$$$            $$$          $$$     $$$     $$$             $$$             $$$$$$$
+     $$$$$         $$$     $$$     $$$$   &&&&   $$$            $$$$$$$$$$$     $$$$$$$$       $$$              $$$$$$$
+     $$$$$         $$$     $$$     $$$$            $$$          $$$     $$$     $$$             $$$             $$$$  $$$$
+     $$$$$         $$$     $$$     $$$$             $$$         $$$     $$$     $$$               $$$           $$$$    $$$$
+     $$$$$         $$$$$$$$$$$     $$$$               $$$$$     $$$     $$$     $$$$$$$$$$          $$$$$       $$$$     $$$$
