+++ src/test/java/com/pricechecker/tui/pricechecker/roomdetails/RoomDetailsParserTest.java
@@ -1,9 +1,11 @@
-package com.pricechecker.tui.pricechecker;
+package com.pricechecker.tui.pricechecker.roomdetails;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.pricechecker.tui.pricechecker.roomdetails.RoomDetails;
+import com.pricechecker.tui.pricechecker.roomdetails.RoomDetailsParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
